+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.7 (2020-06-26)
+
+**Note:** Version bump only for package @arena-im/chat-sdk
+
+
+
+
+
+## 0.1.6 (2020-06-26)
+
+**Note:** Version bump only for package @arena-im/chat-sdk
+
+
+
+
+
 ## 0.1.5 (2020-06-26)
 
 **Note:** Version bump only for package @arena-im/chat-sdk